requirements.txt
flask==1.1.2
mysql.connector==9.1.0